requirements.txt
# ── Backend ──────────────────────────────────────────────────────────────────
fastapi==0.111.0          # REST API framework
uvicorn[standard]==0.29.0 # ASGI server for FastAPI
python-dotenv==1.0.1      # Load .env file for API keys
pydantic==2.7.1           # Request / response validation (bundled with FastAPI)

# ── Google Gemini ─────────────────────────────────────────────────────────────
google-generativeai==0.7.2

# ── Stable Diffusion (CPU) ────────────────────────────────────────────────────
diffusers==0.27.2          # Hugging Face Diffusers (SD pipeline)
transformers==4.41.1       # Tokenizers / model loading
accelerate==0.30.1         # Required by diffusers
torch==2.9.0               # PyTorch CPU build (see install notes below)
torchvision==0.24.0

# ── Image processing ──────────────────────────────────────────────────────────
pillow==10.2.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.32.2
